feat: added the makecoffee method

+++ src/BaseDrink.java
@@ -3,9 +3,9 @@ import java.util.HashMap;
 public class BaseDrink {
     public String name;
     public double cost;
-    public HashMap<String, Double> ingredients = new HashMap<>();
+    public HashMap<String, Integer> ingredients = new HashMap<>();
 
-    public BaseDrink(double water, double milk, double coffee, String name, double cost) {
+    public BaseDrink(int water, int milk, int coffee, String name, double cost) {
         this.name = name;
         this.cost = cost;
         this.ingredients.put("water", water);
